% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.mat.R
\name{create.mat}
\alias{create.mat}
\title{To create and store calculated values of the objective function}
\usage{
create.mat(my_env)
}
\arguments{
\item{my_env}{The environment my_env has various constants stored
from earlier operations dealing with information on the data}
}
\value{
\code{} stores numerical quantities of the objective function
and stores in the two matrices inside the my_env to be accessed by other
functions
}
\description{
This function creates a matrix whose rows and columns depend on the
range or distance of the data and the number of strata solutions that the 
user is seeking to compute. The matrix stores the objective function values
calculated by the algorithm only to be accessed later for the purpose of
presenting the OSB.
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
