% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.res.R
\name{distr.res}
\alias{distr.res}
\title{To organize and print the output to R console}
\usage{
distr.res(my_env)
}
\arguments{
\item{my_env}{My environment my_env has all the solutions stored
in different variables returned by various functions}
}
\value{
\code{} returns the quantities such as OSB, Wh, Vh, WhSh,
nh, Nh and fh
}
\description{
This function is called at the final stage after all OSB and other
associated calculations have been done. It organizes and prints all
the results and outputs to the R console as the final results
}
\examples{
\dontrun{
distr.res()
}

}
\author{
Karuna Reddy <reddy_k@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
