% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.root.R
\name{distr.root}
\alias{distr.root}
\title{Calculate the objective function value for a given (d, y) under a
hypothesized distribution (scaled-data formulation)}
\usage{
distr.root(d, y, c, my_env)
}
\arguments{
\item{d}{numeric: remaining distance/range on the (scaled) axis}

\item{y}{numeric: width of the first stratum on the (scaled) axis}

\item{c}{numeric: per stratum cost multiplier (Ch[k])}

\item{my_env}{environment: holds distribution name, scaled parameters,
and scaling constants ('initval', 'maxval', etc.)}
}
\value{
numeric: sqrt(objective) or -1 if branch is infeasible/invalid
}
\description{
Used by the DP recurrences in the "distribution-known" pathway. All
distribution parameters are interpreted on the "scaled" axis
(initval .. initval+dist), consistent with 'strata.distr()'.
}
\author{
Karuna Reddy <karuna.reddy@auckland.ac.nz>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
