% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erf.R
\name{erf}
\alias{erf}
\title{To calculate the error for a normal variable}
\usage{
erf(x)
}
\arguments{
\item{x}{The data that is provided}
}
\value{
\code{} Gives the error for a normal variable
}
\description{
This function calculates the value of the error according to the          
normally distributed variable using the idea presented in   
Abramowitz and Stegun (2011)
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
