% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hies}
\alias{hies}
\title{Household Income Expenditure Survey (HIES) in Fiji}
\format{
A data frame with 3566 observations on two of
the major quantities collected in the survey. The
variables are:

\describe{
  \item{\code{Expenditure}}{Level of expenditure (FJD)}
  \item{\code{Income}}{Level of income (FJD)}
}
}
\source{
This survey was conducted in 2010 by the Bureau of Statistics (FIBoS) - Fiji Government.
}
\usage{
data(hies)
}
\description{
The hies data comes from the HIES survey conducted
in Fiji in the year 2010. The data contains only two aspects
of the survey.
}
\examples{
data(hies$Income)
min(hies$Income); max(hies$Income)
hist(hies$Income)
boxplot(hies$Income)

}
\keyword{datasets}
