% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math}
\alias{math}
\title{Mathematics Marks for First-year University Students}
\format{
A data frame with 353 observations which represent
mathematics marks and grades for first year math students at
university level. The variable is as follows:

\describe{
\item{\code{cw}}{Coursework marks in 1st year mathematics (0-50)}
\item{\code{end_exam}}{The end of semester examination marks
maths (0-50)}
\item{\code{final_marks}}{Final examination marks in
maths, which is an addition of the cw and end_exam (0-100)}
\item{\code{grade}}{The grade obtained by the student based on the final marks}}
}
\source{
The data was obtained by a masters students at USP, Fiji.
}
\usage{
data(math)
}
\description{
The data contains the mathematics coursework marks, final
examination marks and grades obtained by students in a
first year mathematics course at The University
level in the year 2010 in Fiji.
}
\examples{
data(math)
min(math$final_marks); max(math$final_marks)
hist(math$final_marks)
boxplot(math$final_marks)

}
\keyword{datasets}
