% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minim.val.R
\name{minim.val}
\alias{minim.val}
\title{To identify the minimum value out of two given sets of values}
\usage{
minim.val(val1, val2)
}
\arguments{
\item{val1}{A numeric: the first value}

\item{val2}{A numeric: the second value}
}
\value{
\code{} returns the minimum value
}
\description{
This function is called in data.optim() or distr.optim()
which basically compares and returns the smaller value out
of two given sets of values.
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
