% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realloc.R
\name{realloc}
\alias{realloc}
\title{To re-allocate the stratum sample sizes (nh)}
\usage{
realloc(h, x, nh, Nh, nume, my_env)
}
\arguments{
\item{h}{A numeric: the no. of strata}

\item{x}{A vector: the osb that has been calculated}

\item{nh}{A vector: the stratum sample sizes that have been initially calculated}

\item{Nh}{A vector: the stratum population sizes that have been initially calculated}

\item{nume}{A numeric: the numerator total}

\item{my_env}{The environment my_env has various constants and outputs stored
from earlier opeartions through various other functions}
}
\value{
\code{} calculates and presents the new re-allocate stratum samples
}
\description{
This function re-calculates or re-allocate the stratum sample sizes 
(nh) after it has already been initially allocated via Neyman
allocation. This is applied to resolve the problem of oversampling 
in one or more of the strata.
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
