% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.strata.R
\name{summary.strata}
\alias{summary.strata}
\title{Format and Present Results}
\usage{
\method{summary}{strata}(object, ...)
}
\arguments{
\item{object}{A strata object.}

\item{...}{Additional arguments.

Nicely formatted (and colored) summary for "strata" objects
Console: aligned ASCII table with crayon colors (if supported).
HTML: kable + kableExtra with yellow TOTAL row (text only, no background).}
}
\description{
Format and Present Results
}
