% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.alloc.R
\name{data.alloc}
\alias{data.alloc}
\title{To calculate the stratum sample sizes (nh) for a fixed sample size (n) 
directly based on the data}
\usage{
data.alloc(data, my_env)
}
\arguments{
\item{data}{A vector: provided as an input to the function}

\item{my_env}{The environment my_env has various constants and outputs stored
from earlier opeartions through various other functions}
}
\value{
\code{} calculates and stores quantities such as nh, Nh, Vh, etc.
in the my_env to be accessed and printed as outputs
}
\description{
This function is called towards the final stages of the stratification process
after OSB have been determined. It uses the boundaries to calculate the stratum 
sample allocations using Neyman allocation for all individual strata using the 
raw data.
}
\examples{
\dontrun{
data.alloc()
}

}
\author{
Karuna Reddy <reddy_k@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
