% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.alloc.R
\name{distr.alloc}
\alias{distr.alloc}
\title{To calculate the stratum sample sizes (nh) for a fixed sample size (n) based
on the hypothetical distribution of the data}
\usage{
distr.alloc(my_env)
}
\arguments{
\item{my_env}{The environment my_env which has various constants and outputs stored
from earlier operations through various other functions}
}
\value{
\code{} calculates and stores quantities such as nh, Nh, Vh, etc.
in the my_env to be accessed and printed as outputs
}
\description{
This function is called towards the final stages of the stratification process
after OSB have been determined. It uses the boundaries to calculate the stratum
sample allocations using Neyman allocation for all individual strata using the
underlying distribution of the population.
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr MGM Khan <khan_mg@usp.ac.fj>
}
