% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dist.R
\name{get.dist}
\alias{get.dist}
\title{To identify the best-fit distribution of a univariate data}
\usage{
get.dist(data, my_env)
}
\arguments{
\item{data}{A vector: usually a column in a given data frame}

\item{my_env}{My environment my_env has various constants and data that are
used by the get.dist() function}
}
\value{
\code{} returns a list which contains the best-fit distribution and
its estimated parameters
}
\description{
This function is called at the start of the stratification process where the
best-fit distribution and it parameters are estimated and returned for further
processing towards the computation of stratum boundaries. It basically takes in
the data and fits it with a list of 10 possible distributions and computes
the parameters for all given distributions. It selects the best-fit distribution
to be the one with the lowest AIC
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr MGM Khan <khan_mg@usp.ac.fj>
}
