% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.val.r
\name{mode.val}
\alias{mode.val}
\title{To calculate the modal value of the data}
\usage{
mode.val(x)
}
\arguments{
\item{x}{The data that is provided}
}
\value{
\code{} Gives the mode
}
\description{
This function calculates the value of the mode of the data
that is provided
}
\author{
Karuna Reddy <karuna.reddy@usp.ac.fj>\cr
MGM Khan <khan_mg@usp.ac.fj>
}
